@@ -1,6 +1,6 @@
 __author__ = 'Ichunjo'
 
-__version__ = '0.12.5'
+__version__ = '1.0.0'
 __maintainer__ = 'Ichunjo'
 __email__ = 'ichunjo.le.terrible@gmail.com'
 __status__ = 'Development'
